requirements.txt
streamlit>=1.35.0
streamlit-authenticator>=0.3.2
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0
pyyaml>=6.0
bcrypt>=4.0.0
